% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biomass_tables.R
\name{read_biomass_st1}
\alias{read_biomass_st1}
\title{Summary Table 1 'Natural-origin Pink Salmon' as a tibble.}
\usage{
read_biomass_st1()
}
\value{
tibble with 16 columns of data:
\enumerate{
\item Year
\item Korea
\item Japan
\item M&I
\item WKam
\item EKam
\item WAK
\item SPen
\item Kod
\item CI
\item PWS
\item SEAK
\item NBC
\item SBC
\item WA
\item Total
}
}
\description{
Returns a tibble containing data from Summary Table 1 of
the biomass data, which has abundance (catch and escapement in millions of fish)
of natural-origin Pink Salmon returning to regions of Asia and
North America, 1952-2015.
}
