% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biomass_tables.R
\name{read_biomass_summary_table}
\alias{read_biomass_summary_table}
\title{Summary Table as a tibble from the biomass spreadsheet.}
\usage{
read_biomass_summary_table(sheet, range)
}
\arguments{
\item{sheet}{integer position of the sheet, where 1 is the first sheet.}

\item{range}{string cell range. For example, "A6:P70". This range should include
the row of headings, as well as the data.}
}
\description{
Summary Table as a tibble from the biomass spreadsheet.
}
