% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_biomass_tables.R
\name{read_biomass_st4}
\alias{read_biomass_st4}
\title{Summary Table 4 'Total natural-origin salmon' as a tibble.}
\usage{
read_biomass_st4()
}
\value{
tibble with 5 columns of data:
\enumerate{
\item Pink
\item Chum
\item Sockeye
\item Total
\item Year
}
}
\description{
Returns a tibble containing data from Summary Table 4 of
the biomass data, which has total natural-origin salmon (millions of fish)
returning to regions of Asia and North America, 1952-2015.
}
